--- app/src/main/java/ar/com/wolox/android/cookbook/analytics/AnalyticsEvent.kt
@@ -1,3 +1,0 @@
-package ar.com.wolox.android.cookbook.analytics
-
-abstract class AnalyticsEvent(val name: String, val params: List<Pair<String, String>>)
