+++ app/src/main/java/ar/com/wolox/android/cookbook/analytics/core/AnalyticsEvent.kt
@@ -1,0 +1,3 @@
+package ar.com.wolox.android.cookbook.analytics.core
+
+abstract class AnalyticsEvent(val name: String, val params: List<Pair<String, String>> = emptyList())
